Run the module launcher with new simplified arguments, as a lambda or regular python module:
CcdpModuleLauncher.py is now a single script that handles lambda and non-lambda tasks. CcdpLambdaTaskRunner.py
can now be depracated. A new simplified argument scheme is also provided, with updated help info.

-f specified either a local file or an S3 file in form <bucket_name>:<file_name>. It can be a zip or py file.
If it's a zip file, use -m to specify the mosule.

-m specifies the module name within a zip file (file name minus .py extension). Ignored for non-zip files.

-c specifies a class name. If present, the class is imported from the specified module. Otherwise, it first tries
to invoke the module as a lambda by importing runTask. If that fails, then it runs it as a script.

-o and -r work as before, but are ignored if not an S3 file running as a lambda. There is no way to return a
result when running as a script or invoking a class constructor. If we want to handle results for the latter case,
we can identify a method to call to get the results, and return results if that method is present. I also assumed
no need to save output to files whne not running from an S3 bucket, even for lambda invocation, but I can change that
if needed.

I also updated the logging messages, so a clear progression is captured on the console with the new args scheme